#22 Add safeguards to SerialEntity
Before saving a SerialEntity to the database, a serial must be generated (which can not be done in @PrePersist). This can not be ensured at compile time. A runtime exception is thrown if no serial is set when persisting.
It is also ensured that a serial can only be set once. For IDs that not necessary because they can't be set externally in the first place.
Combined with tests this should be sufficient to ensure the proper use of serials.